+++ libs/src_mock_py/microphysics_scheme_wrapper.py
@@ -19,7 +19,7 @@ wrapper function for an instance of MicrophysicsScheme so it can be used by gene
 and run scripts
 """
 
-from .thermodynamics import Thermodynamics
+from ..thermo import Thermodynamics
 from .mock_microphysics_scheme import MicrophysicsScheme
