+++ scripts/use_microphysics_scheme.py
@@ -8,7 +8,7 @@ Created Date: Tuesday 27th February 2024
 Author: Clara Bayley (CB)
 Additional Contributors:
 -----
-Last Modified: Sunday 1st September 2024
+Last Modified: Monday 2nd September 2024
 Modified By: CB
 -----
 License: BSD 3-Clause "New" or "Revised" License
@@ -23,9 +23,9 @@ import pathlib
 import numpy as np
 
 path = str(pathlib.Path(__file__).parent.resolve())
-sys.path.append(path + "/../libs/")  # add path to src_mock_py to PATH
+sys.path.append(path + "/../libs/")  # add path to python modules to PATH
 
-from src_mock_py.thermodynamics import Thermodynamics
+from thermo.thermodynamics import Thermodynamics
 from src_mock_py.microphysics_scheme_wrapper import MicrophysicsSchemeWrapper
