+++ tests/test_case_1dkid/test_bulkkid.py
@@ -43,7 +43,7 @@ from PyMPDATA_examples import Shipway_and_Hill_2012 as kid
 from PyMPDATA_examples.Shipway_and_Hill_2012 import si
 
 from .perform_1dkid_test_case import perform_1dkid_test_case
-from libs.src_mock_py.thermodynamics import Thermodynamics
+from libs.thermo.thermodynamics import Thermodynamics
 
 
 # TODO(CB): move scheme to seperate file and add documentation
